Add user-level mpd.socket
Starting mpd at login takes a long time, because it scans the entire
library. Sometimes this is even enough to cause the session to fail to
be created properly, triggering a large number of issues.

The mpd package already provides a system-level mpd.socket, but not one
at the user level. This fixes that.