Emacs: Make half-screen scrolling more intuitive
view.el does something weird at the end of the buffer, in that when the
next half-page scroll moves the end of the buffer into view, it
immediately jumps there. This causes unexpected weird and large jumps.

I tried golden-ratio-scroll-screen, and it's pretty nice, but I found
the blinking a bit distracting. An easier way to resolve the question of
where the cursor is on the screen is to just not change it.

scroll-up does have a bit of a weird bug where, near the end of the
buffer, it scrolls one line short of the expected number. This can be
annoying, but it's only really noticeable when you're jumping back and
forth near the end. It's also minor, being only a one line offset,
instead of a massive jump like with view.el. When it happens the cursor
does stay at the appropriate screen position.